requirements.txt
Django>=4.2,<5.0
celery>=5.3
redis>=5.0
flower>=2.0
PyMuPDF>=1.24.0
Pillow>=10.0.0

